server/requirements.txt
Flask==0.12.2
Flask-Script==2.0.5
Flask-PyMongo==0.5.1
python-dotenv==0.7.1